use databse structure
