+++ looker_gen/generator.py
@@ -1,9 +1,6 @@
-from typing import Any, Dict, List
+from typing import Any, Dict, List, Set
 
-from pathlib import Path
-
-from looker_gen.files import FileManager, DBTProject
-from looker_gen.logging import log
+from looker_gen.project import DBTProject
 from looker_gen.types import (
     Dimension,
     DimensionGroup,
@@ -59,10 +56,6 @@ SNOWFLAKE_TYPE_CONVERSIONS = {
 LOOKER_DIM_GROUP_TYPES = ["time", "duration"]
 
 
-def _get_model_name(node_name: NodeName) -> ModelName:
-    return node_name.split(".")[2]
-
-
 # Convert to title case and remove table prefixes
 def _format_label(table_name: str) -> str:
     prefixes = ("dim_", "fct_", "fact_")
@@ -78,61 +71,26 @@ def _format_label(table_name: str) -> str:
 class LookMLGenerator:
     def __init__(self, dbt_dir: str) -> None:
         self.project = DBTProject(dbt_dir)
-        # dbt_path = Path(dbt_dir)
-        # project = FileManager.load_yaml(dbt_dir, "dbt_project.yml")
-        # dbt_target_location = dbt_path.joinpath(project["target-path"])
-        # models_dirs = project.get("model-paths", ["models"])
-
-        # self.project_name = project["name"]
-
-        # self.catalog = FileManager.load_json(dbt_target_location, "catalog.json")
-        # self.manifest = FileManager.load_json(dbt_target_location, "manifest.json")
-
-        # # make column names lower case for lookups; we are not case sensitive
-        # for node_name in self.catalog["nodes"].keys():
-        #     formatted = {
-        #         k.lower(): v for k, v in self.get_catalog_for_node(node_name).items()
-        #     }
-        #     self.catalog["nodes"][node_name]["columns"] = formatted
-
-        # for node_name in self.manifest["nodes"].keys():
-        #     formatted = {
-        #         k.lower(): v for k, v in self.get_manifest_for_node(node_name).items()
-        #     }
-        #     self.manifest["nodes"][node_name]["columns"] = formatted
-
-        # tmp_dir_mapping = FileManager.build_models_dir_mapping(dbt_path, models_dirs)
-        # self.models_dir_mapping = {
-        #     self.get_node_name(k): v for k, v in tmp_dir_mapping.items()
-        # }
-
-        # build explores
         self.explores = self.build_explores()
 
-    def get_model_targets(self, models: str) -> List[str]:
+    def get_model_targets(self, models: str) -> Set[str]:
         if models is None:
-            return [k for k in self.project.catalog["nodes"].keys() if k.startswith(self.project.model_prefix)]
+            return {
+                k
+                for k in self.project.catalog["nodes"].keys()
+                if k.startswith(self.project.model_prefix)
+            }
 
-        return {self.project.get_node_name(m.lower().strip()) for m in models.split(",")}
-
-    # def get_node_name(self, table_name: str) -> NodeName:
-    #     return f"model.{self.project_name}.{table_name}"
-
-    # def get_catalog_for_node(self, node_name: NodeName) -> Dict:
-    #     return self.catalog["nodes"][node_name]["columns"]
-
-    # def get_manifest_for_node(self, node_name: NodeName) -> Dict:
-    #     return self.manifest["nodes"][node_name]["columns"]
+        return {
+            self.project.get_node_name(m.lower().strip()) for m in models.split(",")
+        }
 
     def get_column_config(
         self, node_name: NodeName, column_name: str
     ) -> Dict[str, Any]:
         manifest = self.project.get_manifest_for_node(node_name)
         if column_name in manifest:
-        # if column_name in self.manifest["nodes"][node_name]["columns"]:
-            return manifest[column_name][
-                "meta"
-            ].get("looker-gen", dict())
+            return manifest[column_name]["meta"].get("looker-gen", dict())
 
         return dict()
 
@@ -313,8 +271,9 @@ class LookMLGenerator:
             if self.is_custom_dimension(node_name, col)
         ]
 
-        schema = self.project.catalog["nodes"][node_name]["metadata"]["schema"]
-        table = self.project.catalog["nodes"][node_name]["metadata"]["name"]
+        metadata = self.project.get_catalog_metadata_for_node(node_name)
+        schema = metadata["schema"]
+        table = metadata["name"]
         config = self.get_table_config(node_name)
 
         if "view_label" not in config:
@@ -365,7 +324,7 @@ class LookMLGenerator:
         explores: Dict[str, ExploreConfig] = {}
 
         for node_name in self.project.manifest["nodes"].keys():
-            model_name = _get_model_name(node_name)
+            model_name = self.project.get_model_name(node_name)
             config = self.get_table_config(node_name)
 
             if "explore" in config:
